% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSSM-package.R
\docType{package}
\name{bayesSSM-package}
\alias{bayesSSM}
\alias{bayesSSM-package}
\title{bayesSSM: Bayesian Inference for State-Space Models}
\description{
The bayesSSM package provides implementations of particle filtering,
Particle MCMC, and related methods for Bayesian inference in state-space
models. It includes tools for simulation, posterior inference, and
diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BjarkeHautop/bayesSSM}
  \item \url{https://bjarkehautop.github.io/bayesSSM/}
  \item Report bugs at \url{https://github.com/BjarkeHautop/bayesSSM/issues}
}

}
\author{
\strong{Maintainer}: Bjarke Hautop \email{bjarke.hautop@gmail.com} [copyright holder]

}
\keyword{internal}
