% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.back_transform_params}
\alias{.back_transform_params}
\title{Internal function to back-transform parameters}
\usage{
.back_transform_params(theta_trans, transform)
}
\arguments{
\item{theta_trans}{transformed parameter vector}

\item{transform}{transformation type for each parameter}
}
\value{
original parameter vector
}
\description{
Internal function to back-transform parameters
}
\keyword{internal}
