% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_params_match}
\alias{.check_params_match}
\title{Helper function to validate input of user-defined functions and priors}
\usage{
.check_params_match(
  init_fn,
  transition_fn,
  log_likelihood_fn,
  pilot_init_params,
  log_priors
)
}
\arguments{
\item{init_fn}{A function to initialize the state-space model.}

\item{transition_fn}{A function that defines the state transition of the
state-space model.}

\item{log_likelihood_fn}{A function that calculates the log-likelihood
for the state-space model given latent states.}

\item{pilot_init_params}{A vector of initial parameter values.}

\item{log_priors}{A list of functions for computing the log-prior of each
parameter.}
}
\description{
Helper function to validate input of user-defined functions and priors
}
\keyword{internal}
