% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_log_jacobian}
\alias{.compute_log_jacobian}
\title{Internal function to compute the Jacobian of the transformation}
\usage{
.compute_log_jacobian(theta, transform)
}
\arguments{
\item{theta}{parameter vector (on original scale)}

\item{transform}{transformation type for each parameter}
}
\value{
log-Jacobian of the transformation
}
\description{
Internal function to compute the Jacobian of the transformation
}
\keyword{internal}
