% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ensure_dots}
\alias{.ensure_dots}
\title{Ensure that a function has a `...` argument}
\usage{
.ensure_dots(fun)
}
\arguments{
\item{fun}{A function to modify}
}
\value{
The modified function with `...` added to its formals if it was
not already present.
}
\description{
Ensure that a function has a `...` argument
}
\keyword{internal}
