% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.transform_params}
\alias{.transform_params}
\title{Internal function to transform parameters}
\usage{
.transform_params(theta, transform)
}
\arguments{
\item{theta}{parameter vector}

\item{transform}{transformation type for each parameter}
}
\value{
transformed parameter vector
}
\description{
Internal function to transform parameters
}
\keyword{internal}
