% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS.R
\name{ESS}
\alias{ESS}
\title{Estimate effective sample size (ESS) of MCMC chains.}
\usage{
ESS(chains)
}
\arguments{
\item{chains}{A matrix of dimensions m (iterations) x k (chains).}
}
\value{
The estimated effective sample size (ESS) of the chains.
}
\description{
Estimate effective sample size (ESS) of MCMC chains.
}
\details{
Uses the formula for ESS proposed by Vehtari et al. (2021)
}
\examples{
chains <- matrix(rnorm(3000), nrow = 1000, ncol = 3)
ESS(chains)
}
