% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter-doc.R
\name{particle_filter_returns}
\alias{particle_filter_returns}
\title{Shared Return Values for Particle Filters}
\value{
A list with components:
\describe{
  \item{state_est}{Estimated states over time (weighted mean of particles).}
  \item{ess}{Effective sample size at each time step.}
  \item{loglike}{Total log-likelihood.}
  \item{loglike_history}{Log-likelihood at each time step.}
  \item{algorithm}{The filtering algorithm used.}
  \item{particles_history}{Matrix of particle states over time
  (if \code{return_particles = TRUE}).}
  \item{weights_history}{Matrix of particle weights over time
  (if \code{return_particles = TRUE}).}
}
}
\description{
This block documents the common return value for particle filtering
functions.
}
\keyword{internal}
