% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat}
\alias{rhat}
\title{Compute split Rhat statistic}
\usage{
rhat(chains)
}
\arguments{
\item{chains}{A matrix of dimensions m (iterations) x k (chains).}
}
\value{
The split-Rhat statistic.
}
\description{
Compute split Rhat statistic
}
\details{
Uses the formula for split-Rhat proposed by Gelman et al. (2013).
}
\examples{
chains <- matrix(rnorm(3000), nrow = 1000, ncol = 3)
rhat(chains)
}
\references{
Gelman et al. (2013). Bayesian Data Analysis, 3rd Edition.
}
