% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmh.R
\name{default_tune_control}
\alias{default_tune_control}
\title{Create Tuning Control Parameters}
\usage{
default_tune_control(
  pilot_proposal_sd = 0.5,
  pilot_n = 100,
  pilot_m = 2000,
  pilot_target_var = 1,
  pilot_burn_in = 500,
  pilot_reps = 100,
  pilot_algorithm = c("SISAR", "SISR", "SIS"),
  pilot_resample_fn = c("stratified", "systematic", "multinomial")
)
}
\arguments{
\item{pilot_proposal_sd}{Standard deviation for pilot proposals. Default is
0.5.}

\item{pilot_n}{Number of pilot particles for particle filter. Default is 100.}

\item{pilot_m}{Number of iterations for MCMC. Default is 2000.}

\item{pilot_target_var}{The target variance for the posterior log-likelihood
evaluated at estimated posterior mean. Default is 1.}

\item{pilot_burn_in}{Number of burn-in iterations for MCMC. Default is 500.}

\item{pilot_reps}{Number of times a particle filter is run. Default is 100.}

\item{pilot_algorithm}{The algorithm used for the pilot particle filter.
Default is "SISAR".}

\item{pilot_resample_fn}{The resampling function used for the pilot particle
filter. Default is "stratified".}
}
\value{
A list of tuning control parameters.
}
\description{
This function creates a list of tuning parameters used by the pmmh function.
}
