% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pmmh_output}
\alias{print.pmmh_output}
\title{Print method for PMMH output}
\usage{
\method{print}{pmmh_output}(x, ...)
}
\arguments{
\item{x}{An object of class `pmmh_output`.}

\item{...}{Additional arguments.}
}
\value{
The object `x` invisibly.
}
\description{
Print method for PMMH output
}
\examples{
\dontrun{
print(pmmh_output)
}
}
